Refactored git components and added icons
