+++ app/src/main/java/io/github/jbellis/brokk/gui/Chrome.java
@@ -16,6 +16,9 @@ import io.github.jbellis.brokk.context.FrozenFragment;
 import io.github.jbellis.brokk.git.GitRepo;
 import io.github.jbellis.brokk.gui.dialogs.PreviewImagePanel;
 import io.github.jbellis.brokk.gui.dialogs.PreviewTextPanel;
+import io.github.jbellis.brokk.gui.git.GitIssuesTab;
+import io.github.jbellis.brokk.gui.git.GitPanel;
+import io.github.jbellis.brokk.gui.git.GitPullRequestsTab;
 import io.github.jbellis.brokk.gui.mop.MarkdownOutputPanel;
 import io.github.jbellis.brokk.gui.mop.MarkdownOutputPool;
 import io.github.jbellis.brokk.gui.mop.ThemeColors;
