+++ app/src/main/java/io/github/jbellis/brokk/gui/dialogs/CreatePullRequestDialog.java
@@ -8,8 +8,8 @@ import io.github.jbellis.brokk.git.CommitInfo;
 import io.github.jbellis.brokk.git.GitRepo;
 import io.github.jbellis.brokk.git.GitWorkflowService;
 import io.github.jbellis.brokk.gui.Chrome;
-import io.github.jbellis.brokk.gui.GitCommitBrowserPanel;
 import io.github.jbellis.brokk.gui.components.LoadingButton;
+import io.github.jbellis.brokk.gui.git.GitCommitBrowserPanel;
 import io.github.jbellis.brokk.gui.widgets.FileStatusTable;
 import java.awt.*;
 import java.awt.event.ActionListener;
