+++ app/src/main/java/io/github/jbellis/brokk/gui/util/GitUiUtil.java
@@ -1,4 +1,4 @@
-package io.github.jbellis.brokk.gui;
+package io.github.jbellis.brokk.gui.util;
 
 import com.google.common.base.Splitter;
 import io.github.jbellis.brokk.ContextManager;
@@ -10,6 +10,8 @@ import io.github.jbellis.brokk.difftool.ui.BufferSource;
 import io.github.jbellis.brokk.git.GitRepo;
 import io.github.jbellis.brokk.git.ICommitInfo;
 import io.github.jbellis.brokk.git.IGitRepo;
+import io.github.jbellis.brokk.gui.Chrome;
+import io.github.jbellis.brokk.gui.PrTitleFormatter;
 import io.github.jbellis.brokk.util.SyntaxDetector;
 import java.util.ArrayList;
 import java.util.List;
@@ -128,7 +130,7 @@ public final class GitUiUtil {
                 var commitContent = repo.getFileContent(commitId, file);
 
                 SwingUtilities.invokeLater(() -> {
-                    var brokkDiffPanel = new BrokkDiffPanel.Builder(chrome.themeManager, cm)
+                    var brokkDiffPanel = new BrokkDiffPanel.Builder(chrome.getTheme(), cm)
                             .leftSource(new BufferSource.StringSource(parentContent, parentCommitId, file.toString()))
                             .rightSource(new BufferSource.StringSource(commitContent, commitId, file.toString()))
                             .build();
@@ -330,7 +332,7 @@ public final class GitUiUtil {
                 String finalDialogTitle = "Diff: %s [Local vs %s]".formatted(file.getFileName(), baseCommitShort);
 
                 SwingUtilities.invokeLater(() -> {
-                    var brokkDiffPanel = new BrokkDiffPanel.Builder(chrome.themeManager, cm)
+                    var brokkDiffPanel = new BrokkDiffPanel.Builder(chrome.getTheme(), cm)
                             .leftSource(new BufferSource.StringSource(
                                     finalOldContent, finalBaseCommitTitle, file.toString()))
                             .rightSource(
@@ -467,7 +469,7 @@ public final class GitUiUtil {
                     return;
                 }
 
-                var builder = new BrokkDiffPanel.Builder(chrome.themeManager, cm);
+                var builder = new BrokkDiffPanel.Builder(chrome.getTheme(), cm);
                 var parentId = commitInfo.id() + "^";
 
                 for (var file : files) {
@@ -505,7 +507,7 @@ public final class GitUiUtil {
                     return;
                 }
 
-                var builder = new BrokkDiffPanel.Builder(chrome.themeManager, cm);
+                var builder = new BrokkDiffPanel.Builder(chrome.getTheme(), cm);
                 var parentId = commitInfo.id() + "^";
 
                 // Track target file index
@@ -560,7 +562,7 @@ public final class GitUiUtil {
                     return;
                 }
 
-                var builder = new BrokkDiffPanel.Builder(chrome.themeManager, contextManager);
+                var builder = new BrokkDiffPanel.Builder(chrome.getTheme(), contextManager);
                 var repo = contextManager.getProject().getRepo();
                 var shortId = shortenCommitId(commitInfo.id());
